feat: DefaultDatabaseService 정의 및 더미 코드 이동
