+++ Clipster/Clipster/Data/Protocol/Network/DatabaseService.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+protocol DatabaseService {
+    func dummyFetchFolders() async -> [FolderDTO]?
+    func dummyDownloadScreenshot() async -> Data?
+}
